requirements.txt
mysql-connector-python
playsound==1.2.2
